# server: optimize bce on the packetio.writePacket