Signed-off-by: Weizhen Wang <wangweizhen@pingcap.com>

<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Go is a memory safe language. In array/slice element indexing and subslice operations, Go runtime will check whether or not the involved indexes are out of range. If an index is out of range, a panic will be produced to prevent the invalid index from doing harm. This is called bounds check. Bounds checks make our code run safely, on the other hand, they also make our code run a little slower.

Issue Number: ref #34669

Problem Summary:

### What is changed and how it works?

before 

```
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/server
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkPacketIOWrite-16    	10000000	       871.3 ns/op
PASS
ok  	github.com/pingcap/tidb/server	9.875s
```

after

```
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/server
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkPacketIOWrite-16    	10000000	       830.0 ns/op
PASS
ok  	github.com/pingcap/tidb/server	9.102s
```

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```